renamed back svm-ts directory to svm-js
