+++ src/bp/ml/svm-js/core/config.ts
