+++ src/bp/ml/svm-js/core/default-config.ts
