+++ src/bp/ml/svm-js/core/kernel-types.ts
