+++ src/bp/ml/svm-js/core/svm-types.ts
