+++ src/bp/ml/svm-js/core/svm.ts
