+++ src/bp/ml/svm-js/util/average.ts
