+++ src/bp/ml/svm-js/util/cross-combinations.ts
