+++ src/bp/ml/svm-js/util/normalize-input.ts
