+++ src/bp/ml/svm-js/util/reduce-dataset.ts
