+++ src/bp/ml/svm-js/util/standard-deviation.ts
